added screen to texture in demo + CAPI method

+++ BonEngine/inc/Assets/Types/Image.h
@@ -65,7 +65,7 @@ namespace bon
 			 *
 			 * \param handle Image handle.
 			 */
-			_Image(_ImageHandle* handle) : IAsset(std::to_string((long)handle).c_str()) {
+			_Image(_ImageHandle* handle) : IAsset(std::to_string((long long)handle).c_str()) {
 				_SetHandle(handle);
 			}
